+++ qt/gen_qaccessibleplugin.go
@@ -65,7 +65,7 @@ func QAccessiblePlugin_TrUtf8(s string) string {
 func (this *QAccessiblePlugin) Create(key string, object *QObject) *QAccessibleInterface {
 	key_Cstring := C.CString(key)
 	defer C.free(unsafe.Pointer(key_Cstring))
-	ret := C.QAccessiblePlugin_Create(this.h, key_Cstring, C.ulong(len(key)), object.cPointer())
+	ret := C.QAccessiblePlugin_Create(this.h, key_Cstring, C.size_t(len(key)), object.cPointer())
 	return newQAccessibleInterface_U(unsafe.Pointer(ret))
 }
